requirements.txt
opencv-python
mediapipe
pywin32
pystray
Pillow
